Cleaned up routes and rewrote some controllers for increased accuracy and safety.

+++ app/controllers/call.controller.js
@@ -1,16 +1,17 @@
 const db = require("../models");
 
 const User = db.user;
-const Group = db.group;
 const Call = db.call; 
+const Message = db.message;
 
-exports.createCall = (req, res) => {
+exports.createCall = (req, res) => { 
     Call.create({
         title: req.body.title,
         location: req.body.location,
         time: req.body.time,
         contactinfo: req.body.contactinfo,
         description: req.body.description,
+        group: req.body.group,
         image: req.body.image
     })
         .then(call => {
@@ -19,25 +20,9 @@ exports.createCall = (req, res) => {
                     id: req.params.id
                 }
             }).then(user => {
-                if (!user) {
-                    Group.findOne({
-                        where: {
-                            id: req.params.id
-                        }
-                    }).then(group => {
-                        if (!group) {
-                            res.send({ message: 'No such User or Group' });
-                        } else {
-                            group.addCalls(call).then(() => {
-                                res.send({ message: `${group.title} called for ${call.title}` });
-                            })
-                        }
-                    })
-                } else {
-                    user.addCalls(call).then(() => {
-                        res.send({ message: `${user.username} called for ${call.title}` });
-                    })
-                }
+                user.addCalls(call).then(() => {
+                    res.send({ message: `${user.username} called for ${call.title}` });
+                })
             })
         })
         .catch(err => {
@@ -46,14 +31,26 @@ exports.createCall = (req, res) => {
 };
 
 exports.deleteCall = (req, res) => {
-    Call.destroy({
+    Call.findOne({
         where: {
             id: req.body.callId
         }
     })
-      .then(
-          res.send({ message: 'Call was deleted.' })
-      )
+      .then(call => {
+          call.getUsers().then(users => {
+              if (users.some(user => user.id === req.params.id)) {
+                  Call.destroy({
+                      where: {
+                          id: req.body.callId
+                      }
+                  }).then(() => {
+                    res.send({ message: 'Call deleted.' });
+                  })
+              } else {
+                  res.send({ message: 'Users can only delete their own calls.' });
+              }
+          })
+      })
       .catch(err => {
         res.status(500).send({ message: err.message });
       });
@@ -62,37 +59,12 @@ exports.deleteCall = (req, res) => {
 exports.updateCall = (req, res) => {
     Call.findOne({
         where: {
-            id: req.body.callId
+            id: req.body.callId 
         }
     })
         .then(call => {
-            User.findOne({
-                where: {
-                    id: req.params.id
-                }
-            }).then(user => {
-                if (!user) {
-                    Group.findOne({
-                        where: {
-                            id: req.params.id
-                        }
-                    }).then(group => {
-                        if (!group) {
-                            res.send({ message: 'No such User or Group' });
-                        } else {
-                            call.update({
-                                title: req.body.title,
-                                location: req.body.location,
-                                time: req.body.time,
-                                contactinfo: req.body.contactinfo,
-                                description: req.body.description,
-                                image: req.body.image
-                            }).then(() => {
-                               res.send({ message: `${call.title} has been updated by ${group.title}` });
-                            })
-                        }
-                    })
-                } else {
+            call.getUsers().then(users => {
+                if (users.some(user => user.id === req.params.id)) {
                     call.update({
                         title: req.body.title,
                         location: req.body.location,
@@ -101,8 +73,10 @@ exports.updateCall = (req, res) => {
                         description: req.body.description,
                         image: req.body.image
                     }).then(() => {
-                       res.send({ message: `${call.title} has been updated by ${user.username}` });
+                       res.send({ message: `${call.title} has been updated.` });
                     })
+                } else {
+                    res.send({ message: 'Users can only update their own calls.' });
                 }
             })
         })
@@ -135,52 +109,17 @@ exports.getOneCall = (req, res) => {
         });
 };
 
-exports.getCallsOf = (req, res) => {
-    User.findOne({
+exports.getCallMessages = (req, res) => {
+    Message.findAll({
         where: {
-            id: req.params.id
-        }
-    }).then(user => {
-        if (!user) {
-            Group.findOne({
-                where: {
-                    id: req.params.id
-                }
-            }).then(group => {
-                if (!group) {
-                    res.send({ message: 'No such User or Group' });
-                } else {
-                  var groupcalls = [];
-                  group.getCalls().then(calls => {
-                    for (let i = 0; i < calls.length; i++) {
-                      groupcalls.push({
-                          callId: calls[i].id,
-                          title: calls[i].title
-                        });
-                    }
-                    res.status(200).send({ 
-                      groupcalls: groupcalls
-                    });
-                  });
-                }
-            })
-        } else {
-          var usercalls = [];
-          user.getCalls().then(calls => {
-            for (let i = 0; i < calls.length; i++) {
-              usercalls.push({
-                  callId: calls[i].id,
-                  title: calls[i].title
-                });
-            }
-            res.status(200).send({ 
-              usercalls: usercalls
-            });
-          });
+            otherId: req.params.id
         }
     })
-    .catch(err => {
+      .then(messages => {
+        res.send(messages);
+      })
+      .catch(err => {
         res.status(500).send({ message: err.message });
-    });
+      });
 };
